backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic-settings>=2.6.0
itsdangerous>=2.2.0
google-genai>=1.0.0
requests>=2.31.0

